Add brcypt gem & fix graphiql
This commit is the result of experimenting with GraphQL.
I was not able to create my first User model because of
some errors with GraphQL (more details when I'll find a
fix).

This is the first commit to try to clean up the databse
to go back to something working with graphiql.